consolidate api fetching methods
