+++ store/getMeta.ts
@@ -5,7 +5,7 @@ import { exec } from 'child_process';
 import { promisify } from 'util';
 import redis from './redis';
 import { buildReplayUrl, redisCount } from '../util/utility';
-import { readGcData } from './queries';
+import { readGcData } from './getGcData';
 const execPromise = promisify(exec);
 
 // Get a sample meta file
@@ -18,7 +18,6 @@ const builder = root.loadSync('./proto/dota_match_metadata.proto', {
 const Message = builder.lookupType('CDOTAMatchMetadataFile');
 
 export async function getMeta(matchId: string) {
-  // Check if we have the info to construct url
   const gcdata = await readGcData(Number(matchId));
   if (!gcdata) {
     return null;
